+++ backend/__init__.py
@@ -25,6 +25,8 @@ from api.projects import projects
 from api.users import users
 from api.comments import comments
 from api.bookmarks import bookmarks
+from api.avatars import avatars
+
 from api.search import search, prepare_search_results
 from api.helper.apiexception import ApiException
 from api.helper.images import Identicon
@@ -337,6 +339,8 @@ app.register_blueprint(users)
 app.register_blueprint(search)
 app.register_blueprint(comments)
 app.register_blueprint(bookmarks)
+app.register_blueprint(avatars)
+
 
 
 if __name__ == "__main__":
